auth: move CborReader into its own header
Extract the internal CBOR decoder from Auth.cpp into a self-contained
header-only src/auth/CborReader.h (namespace openmoq::moqx::auth). Auth.cpp
now includes it; the claim-parsing helpers are unchanged. No build wiring
needed since moqx_core picks up headers via the src/ include dir.
